+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Update Event API to support both RES 1.0 and 0.42+. ([@palkan][])
+
 ## 0.1.0 (2020-04-22)
 
 - Open source Active Event Store. ([@palkan][])